pyproject.toml
[tool.poetry]
name = "telegram-finance-bot"
version = "0.1.0"
description = ""
authors = ["Mikhail Tarbeev <tarbeev2003@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.6"
aiogram = "2.4"
pytz = "^2023.3.post1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
